self-affinities no longer influence calculations
Previously, self affinity was included in the mean
affinity of an element to it's own cluster.

With self affinitly usually set to 1, since an
object should be perfectly similar with itself,
small clusters had a boosted affinity within
the clusters equal to
(1/n)(1-avg_without_self_affinity)

The effect was to protect small clusters (<4 elems)
from being absorbed into other clusters, even
when the affinity between the elements in the
cluster was weak.

The case of a single element cluster is
handled by assigning an affinity to it's
own cluster equal to \alpha. It will not
be disbanded by default, but it will
easily move to any other cluster with
an affinity greater than \alpha, and
the empty cluster can be removed.