Allow plotting to use log scales if average similarity is low
On a linear scale, anything less than 0.1 looks mostly black.
Average similarity can easily drop below 0.01, and appropriate
thresholds for clustering can also be below 0.01.
Therefore, on a linear scale, it appears that a lot of black ends up
inside clusters.

On the pseudo-log scale, 0 remains black, but very small similarities are
distinguishable.

The approach uses a threshold parameter, as linear scale is still
appropriate when the mean similarity is >0.1.

Code is opinionated, but threshold can be tuned if needed.